revise: gate entry point functions behind cli mod
