+++ wdl/Cargo.toml
@@ -41,9 +41,9 @@ codespan-reporting = { workspace = true }
 
 [features]
 default = ["ast", "grammar", "lint", "format", "engine"]
-analysis = ["dep:wdl-analysis", "dep:anyhow", "dep:indicatif", "dep:tokio", "dep:url"]
+analysis = ["dep:wdl-analysis"]
 ast = ["dep:wdl-ast"]
-engine = ["analysis", "dep:wdl-engine", "dep:codespan-reporting", "dep:serde_json"]
+engine = ["dep:wdl-engine"]
 format = ["dep:wdl-format"]
 doc = ["dep:wdl-doc"]
 grammar = ["dep:wdl-grammar"]
