+++ wdl/src/bin/wdl.rs
@@ -24,13 +24,13 @@ use codespan_reporting::term::termcolor::ColorChoice;
 use codespan_reporting::term::termcolor::StandardStream;
 use colored::Colorize;
 use tracing_log::AsTrace;
-use wdl::analyze;
 use wdl::ast::Diagnostic;
 use wdl::ast::Document;
 use wdl::ast::Validator;
+use wdl::cli::analyze;
+use wdl::cli::run;
+use wdl::cli::validate_inputs;
 use wdl::lint::LintVisitor;
-use wdl::run;
-use wdl::validate_inputs;
 use wdl_ast::Node;
 use wdl_ast::Severity;
 use wdl_doc::document_workspace;
